# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PAJOY Smart Business System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PRODUCTS ────────────────────────────────────────────────────────────────
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: q
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}/purchases:
    get:
      operationId: getCustomerPurchases
      tags: [customers]
      summary: Get customer purchase history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"

  # ─── SUPPLIERS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ─── USER ─────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role, is_active]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, admin, manager, cashier, production_staff, accountant]
        branch_id:
          type: ["integer", "null"]
        is_active:
          type: boolean
        avatar_url:
          type: ["string", "null"]
        last_login:
          type: ["string", "null"]
        created_at:
          type: string
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        parent_id:
          type: ["integer", "null"]
        sort_order:
          type: ["integer", "null"]
        is_active:
          type: boolean
        product_count:
          type: ["integer", "null"]
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        parent_id:
          type: ["integer", "null"]
        sort_order:
          type: ["integer", "null"]
        is_active:
          type: boolean

    # ─── PRODUCT ──────────────────────────────────────────────────────────────
...
    Customer:
      type: object
      required: [id, name, phone]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: string
        alt_phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        id_number:
          type: ["string", "null"]
        customer_type:
          type: ["string", "null"]
        credit_limit:
          type: ["number", "null"]
        balance:
          type: number
        loyalty_points:
          type: integer
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
        total_purchases:
          type: ["number", "null"]
        purchase_count:
          type: ["integer", "null"]
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: string
        alt_phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        id_number:
          type: ["string", "null"]
        customer_type:
          type: ["string", "null"]
        credit_limit:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    CustomerListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── SUPPLIER ─────────────────────────────────────────────────────────────
...
    SaleItem:
      type: object
      required: [product_id, product_name, quantity, unit_price, total]
      properties:
        product_id:
          type: integer
        variant_id:
          type: ["integer", "null"]
        product_name:
          type: string
        sku:
          type: ["string", "null"]
        quantity:
          type: integer
        unit_price:
          type: number
        cost_price:
          type: ["number", "null"]
        discount:
          type: number
        total:
          type: number
    Sale:
      type: object
      required: [id, sale_number, subtotal, total, payment_method, payment_status, created_at]
      properties:
        id:
          type: integer
        sale_number:
          type: string
        customer_id:
          type: ["integer", "null"]
        customer_name:
          type: ["string", "null"]
        cashier_id:
          type: ["integer", "null"]
        cashier_name:
          type: ["string", "null"]
        subtotal:
          type: number
        discount_amount:
          type: number
        tax_amount:
          type: number
        total:
          type: number
        amount_paid:
          type: number
        change_given:
          type: number
        payment_method:
          type: string
        payment_status:
          type: string
        voided:
          type: boolean
        void_reason:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        created_at:
          type: string
...